Control access was created over the methods in each ticket request

+++ src/controllers/ticket.controller.js
@@ -34,8 +34,14 @@ export const createTicket = async (req, res) => {
 
 export const getTicket = async (req, res) => {
   try {
-    const ticket = await Ticket.findById(req.params.id).populate("user");
+    const ticket = await Ticket.findById(req.params.id).populate([
+      "user",
+      "receiver",
+    ]);
+    const userId = req.user.id;
     if (!ticket) return res.status(404).json({ message: "Ticket not found" });
+    if (!(userId === ticket.user.id || userId === ticket.receiver.id))
+      return res.status(403).json({ message: "Access denied" });
     res.json(ticket);
   } catch (error) {
     return res.status(404).json({ message: "Ticket not found" });
@@ -44,8 +50,18 @@ export const getTicket = async (req, res) => {
 
 export const deleteTicket = async (req, res) => {
   try {
-    const ticket = await Ticket.findByIdAndDelete(req.params.id);
+    const ticket = await Ticket.findById(req.params.id);
     if (!ticket) return res.status(404).json({ message: "Ticket not found" });
+    if (
+      !(
+        req.user.id === ticket.user.toString() ||
+        req.user.id === ticket.receiver.toString()
+      )
+    )
+      return res.status(403).json({ message: "Access denied" });
+    const deletedTicket = await Ticket.findByIdAndDelete(req.params.id);
+    if (!deletedTicket)
+      return res.status(404).json({ message: "Ticket not found" });
     return res.sendStatus(204);
   } catch (error) {
     return res.status(404).json({ messege: "Ticket not found" });
@@ -54,11 +70,25 @@ export const deleteTicket = async (req, res) => {
 
 export const updateTicket = async (req, res) => {
   try {
-    const ticket = await Ticket.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-    });
-    if (!ticket) return res.status(404).json({ messege: "Ticket not found" });
-    res.json(ticket);
+    const ticket = await Ticket.findById(req.params.id);
+    if (!ticket) return res.status(404).json({ message: "Ticket not found" });
+    if (
+      !(
+        req.user.id === ticket.user.toString() ||
+        req.user.id === ticket.receiver.toString()
+      )
+    )
+      return res.status(403).json({ message: "Access denied" });
+    const updateTicket = await Ticket.findByIdAndUpdate(
+      req.params.id,
+      req.body,
+      {
+        new: true,
+      }
+    );
+    if (!updateTicket)
+      return res.status(404).json({ message: "Ticket not found" });
+    res.json(updateTicket);
   } catch (error) {
     return res.status(404).json({ messege: "Ticket not found" });
   }
